refactor(edition): types from lib
helps with no unchecked indexed access
