+++ dotcom-rendering/src/amp/components/Blocks.tsx
@@ -4,7 +4,8 @@ import React from 'react';
 import { pillarPalette_DO_NOT_USE } from '../../lib/pillars';
 import type { CommercialProperties } from '../../types/commercial';
 import type { Switches } from '../../types/config';
-import type { EditionId } from '../../types/edition';
+import { EditionId } from '../../web/lib/edition';
+
 import { blockLink } from '../lib/block-link';
 import { findBlockAdSlots } from '../lib/find-adslots';
 import { Elements } from './Elements';
