+++ dotcom-rendering/src/model/window-guardian.ts
@@ -1,5 +1,5 @@
 import type { ConfigType, ServerSideTests, Switches } from '../types/config';
-import type { EditionId } from '../types/edition';
+import { EditionId } from '../web/lib/edition';
 
 export interface WindowGuardianConfig {
 	isDotcomRendering: boolean;
