+++ dotcom-rendering/src/types/front.ts
@@ -1,6 +1,6 @@
 import type { ArticlePillar, ArticleSpecial } from '@guardian/libs';
+import type { EditionId } from '../web/lib/edition';
 import type { ServerSideTests, Switches } from './config';
-import type { EditionId } from './edition';
 import type { FooterType } from './footer';
 import type { FETagType } from './tag';
 import type { TrailType } from './trails';
