+++ dotcom-rendering/src/types/frontend.ts
@@ -1,7 +1,7 @@
+import type { EditionId } from '../web/lib/edition';
 import type { BadgeType } from './badge';
 import type { CommercialProperties } from './commercial';
 import type { ConfigType } from './config';
-import type { EditionId } from './edition';
 import type { FooterType } from './footer';
 import type { CAPIOnwards } from './onwards';
 import type { TagType } from './tag';
