+++ dotcom-rendering/src/web/components/ContainerTitle.tsx
@@ -6,7 +6,7 @@ import {
 	space,
 	text,
 } from '@guardian/source-foundations';
-import type { EditionId } from '../../types/edition';
+import type { EditionId } from '../lib/edition';
 import type { DCRContainerPalette } from '../../types/front';
 import type { Colour } from '../../types/palette';
 import { decideContainerOverrides } from '../lib/decideContainerOverrides';
