+++ dotcom-rendering/src/web/components/Footer.tsx
@@ -18,7 +18,7 @@ import {
 } from '@guardian/source-react-components';
 import { clearFix } from '../../lib/mixins';
 import type { PillarType } from '../../model/extract-nav';
-import type { EditionId } from '../../types/edition';
+import type { EditionId } from '../lib/edition';
 import type { FooterType } from '../../types/footer';
 import { BackToTop } from './BackToTop';
 import { Island } from './Island';
