+++ dotcom-rendering/src/web/components/Nav/Nav.tsx
@@ -8,7 +8,8 @@ import {
 } from '@guardian/source-react-components';
 import { clearFix } from '../../../lib/mixins';
 import type { NavType } from '../../../model/extract-nav';
-import type { EditionId } from '../../../types/edition';
+import { EditionId } from '../../lib/edition';
+
 import { GuardianRoundel } from '../GuardianRoundel';
 import { Hide } from '../Hide';
 import { Pillars } from '../Pillars';
